make nsUptimeMillis be uint32 on armv6
this means that millis will wrap in ~49 days, and nano in about 1 hour
which isn't ideal, but I can't see how else to deal with it.
AFAICT none of the in-tree users of renderscript use nano, and the
users of millis aren't affected by it wrapping.

(cherry picked from commit 2e5b9def974ef8207c3db7fcc81fdce45e9dabd7)

Change-Id: I724d9d7edf4bd5d10ec94234a33a3f9dd5074d4b